pyproject.toml
[tool.poetry]
name = "wall-e"
version = "0.1.0"
description = ""
authors = ["Chuck Conway <charles@cconway.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi = {extras = ["all"], version = "^0.110.2"}
pydantic = "^2.7.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
install = "^1.3.5"
python-dotenv = "^1.0.1"
prometheus-fastapi-instrumentator = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
